Move user and register
Took 1 minute
